pyproject.toml
[project]
name = "pp-2-0"
version = "0.1.0"
description = "Wagon Classification API with PostgreSQL"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.8",
    "httpx>=0.28.1",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "opencv-python>=4.13.0.92",
    "pandas>=2.3.3",
    "patool>=1.15.0",
    "pillow>=11.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.3",
    "unrar>=0.4",
    "uvicorn[standard]>=0.39.0",
    "winrar>=0.0.1",
    "prometheus-client>=0.20.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.23",
    "bcrypt>=4.1.0",
]

# ============================================
# НАСТРОЙКИ ДЛЯ ИНСТРУМЕНТОВ (Python 3.9)
# ============================================

[tool.black]
line-length = 88
target-version = ['py39']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | scripts
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["scripts", "migrations"]
skip_glob = ["**/__pycache__/*"]
py_version = 39

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv", "venv", "scripts", "migrations"]
per-file-ignores = ["__init__.py:F401"]

[tool.pylint.main]
ignore = ["scripts", "migrations", "tests"]
fail-under = 7
py-version = "3.9"

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "R0903",  # too-few-public-methods
    "W0707",  # raise-missing-from
    "W1203",  # logging-fstring-interpolation
    "R0801",  # duplicate-code
    "E0401",  # import-error
    "C0413",  # wrong-import-position
    "C0411",  # wrong-import-order
    "W0611",  # unused-import
    "W1309",  # f-string-without-interpolation
    "W0718",  # broad-exception-caught
    "C0103",  # invalid-name
]

[tool.pylint.design]
max-args = 7
max-locals = 20
max-attributes = 15

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
exclude = ["scripts", "tests", "migrations"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# ============================================
# НАСТРОЙКИ ДЛЯ COVERAGE (ИСКЛЮЧЕНИЯ)
# ============================================

[tool.coverage.run]
source = ["app"]
omit = [
    # __init__.py файлы (обычно пустые)
    "app/__init__.py",
    "app/domain/__init__.py",
    "app/infrastructure/__init__.py",
    "app/infrastructure/database/__init__.py",
    "app/presentation/__init__.py",
    "app/presentation/api/__init__.py",
    "app/presentation/cli/__init__.py",
    "app/use_cases/__init__.py",
    
    # Скрипты и утилиты (не требуют тестов)
    "app/infrastructure/archive_extractor.py",
    "app/infrastructure/augmentation_service.py",
    "app/infrastructure/user_repository.py",
    "app/infrastructure/logger.py",
    "app/presentation/cli/train_cli.py",
    "app/use_cases/train_model.py",
    
    # Требуют реальной БД (интеграционные тесты)
    "app/infrastructure/database/connection.py",
    "app/infrastructure/database/models.py",
    "app/infrastructure/database/postgres_user_repository.py",
    "app/main.py",
    "app/presentation/api/routes.py",
    "app/presentation/api/dependencies.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass",
    "import abc",
    "from abc import ABC",
    "@abstractmethod",
]